Lots of changes
Would have committed this earlier, but I forgot that it was already in
git.  It was almost working, but was not preserving the file timestamps.
I'm about to switch to apache IO utils for the file copy, so copying is
temporarily disabled.  Also, a progress bar has been added, and is
almost completely hooked up.